fix: use current_schema() for Connection#getSchema
Previously the driver used "show search_path", however it might result into non-evaluated '$user' value

fixes #356
